For v5 define a version of the enc_len function.
In version 5 enc_len was renamed to onig_enc_len and an extra arg
was added. Its not currently clear what the extra arg does, as its
ignored when the macro is expanded. We've defined our enc_len macro
in terms of the ONIGENC_MBC_ENC_LEN() which onig_enc_len uses.